Use composite_primary_keys id formatting
ActiveRecord's implementation od compose primary keys has its own way to
serialize the combined identifier into a single string to use as a parameter
or to pass to find. We want to continue to use the version
composite_primary_key expects to override to_param to return the "old" version.