Disable active_record validation of composite primary key declaration
With its implementation of composite primary keys, active_record adds a
validation method to make sure you are declaring the composite primary key the
way it expects.  The composite_primary_keys gem expects you to declare the key
differently, so it always fails the validation. The simplest workaround is to
turn the validation off with a noop.